revert topology
